Story/requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.12.0
fpdf2>=2.7.0
ebooklib>=0.18
schedule>=1.2.0
